\name{BANOVAoutput}

\alias{BANOVAoutput}

\title{Output of ANOVA function}

\description{
  We provides several txt outputs.
  1.ANOVA stat.txt                                                             
    -->ANOVA:Cmax, AUC0t, AUC0inf, ln(Cmax), ln(AUC0t), ln(AUC0inf)         
    -->90CI: ln(Cmax), ln(AUC0t), and ln(AUC0inf)                          
}

\keyword{misc}