\name{BANOVAplot}

\alias{BANOVAplot}

\title{BANOVAplot}


\description{
  We provides several pdf. outputs about BANOVA plots for outlier detections.
    1.Normal Probability Plot of lnCmax (intrasubj)
    2.Normal Probability Plot of lnCmax (intersubj)
    3.lnCmax(expected value) vs. studentized residuals(intrasubj)
    4.lnCmax(expected value) vs. studentized residuals(intersubj)
    5.Normal Probability Plot of lnAUC0t (intrasubj)
    6.Normal Probability Plot of lnAUC0t (intersubj)
    7.lnAUC0t(expected value) vs. studentized residuals(intrasubj)
    8.lnAUC0t(expected value) vs. studentized residuals(intersubj)
    9.Normal Probability Plot of lnAUC0INF (intrasubj)
    10.Normal Probability Plot of lnAUC0INF (intersubj)
    11.lnAUC0INF(expected value) vs. studentized residuals(intrasubj)
    12.lnAUC0INF(expected value) vs. studentized residuals(intersubj)
}

\keyword{misc}