\encoding{UTF-8}

\name{bear.setup}

\alias{bear.setup}

\title{Setup module for bear}

\description{
  You can setup the method/value for:
   1. lambda_z estimation,
   2. AUC calculation,
   3. BE criterion (LL)
   4. Dose,
   5. Dosing Interval,
   6. Tlast, and
   7. x-axis & y-axis labels for plots.
}

\keyword{misc}