\encoding{UTF-8}

\name{create.products_sum}

\alias{create.products_sum}

\title{create.products_sum}

\description{
this function is to create two temporary .csv files for
the list of the section containing the means (SD) for each formulation 
(by time point) in NCA output file. These two .csv files will be deleted
after nca_outputs file is done.
}

\keyword{misc}