\encoding{UTF-8}

\name{demoBANOVA}

\alias{demoBANOVA}

\title{Statistical analysis (ANOVA(lm), 90CI...)for demo file}

\description{
With a two-treatment, two-period, two-sequence randomized crossover
design, ANOVA statistical model includes factors of the following sources:
sequence, subjects nested in sequences, period and treatment. ANOVA will
be applied to obtain estimates for the adjusted differences between
treatment means and the standard error associated with these differences.
Log-transformed BA measures will also be analyzed.
}

\keyword{misc}