\name{demomenu1}

\alias{demomenu1}

\title{menu for NCA demo file}

\description{
lambda z est. from the exact 3 data points--> Statistical analysis,
lambda z est. with adjusted R sq. (ARS)--> Statistical analysis,
lambda z est. with Akaike information criterion (AIC)--> Statistical analysis,
lambda z est. with Two-Times-Tmax method (TTT)--> Statistical analysis,
lambda z est. with TTT and ARS--> Statistical analysis,
lambda z est. with TTT and AIC--> Statistical analysis,
}

\keyword{misc}