\encoding{UTF-8}

\name{description_drugcode}

\alias{description_drugcode}

\title{Description for drug code}

\description{
Data Codes:
 Drug:
     1: Ref.
     2: Test
 Sequence:
     1: Ref. --> Test
     2: Test --> Ref.
 Period:
     1: 1st-treatment period
     2: 2nd-treatment period
}

\keyword{misc}