\name{description_NCAinput}

\alias{description_NCAinput}

\title{Description for NCA input data}

\description{

Input/Edit Data
   -> subject no.(subj)
   -> sequence (seq)
         1:Ref.-->Test
         2:Test-->Ref.
   -> period (prd)
         1: 1st-treatment period
         2: 2nd-treatment period
   -> sampling time
   -> drug plasma/serum/blood concentration (conc)
}

\keyword{misc}