\encoding{UTF-8}

\name{description_ParaNCAcsv}

\alias{description_ParaNCAcsv}

\title{Description for NCA csv file for parallel study}

\description{

Data file should consist of
    row1: column title, such as subj, drug, time, conc & etc.
 column1: subject no.(subj)
 column2: treatment (drug)
 column3: sampling time
 column4: drug plasma/serum/blood concentration (conc)

}

\keyword{misc}