\encoding{UTF-8}

\name{description_ParaNCAinput}

\alias{description_ParaNCAinput}

\title{Description for NCA input data for parallel study}

\description{

Input/Edit Data
   -> subject no.(subj)
   -> treatment (drug)
          1: Ref.
          2: Test
   -> sampling time
   -> drug plasma/serum/blood concentration (conc)
}

\keyword{misc}