\encoding{UTF-8}

\name{description_plot}

\alias{description_plot}

\title{Description for plot}

\description{

Authors: Hsin-ya Lee, Yung-jin Lee
Kaohsiung Veterans General Hospital (HY) &
ptpc inc. (YJ),
Kaoshiung, Taiwan 80794
E-mail: mobilepk@gmail.com
website: http://pkpd.kmu.edu.tw/bear

}

\keyword{misc}