\encoding{UTF-8}

\name{description_RepMIX}

\alias{description_RepMIX}

\title{Description for lme for replicated study}

\description{
With a two-treatment, more than two period, two-sequence randomized
crossover design, linear mixed-effects model may be used.  The lme in nlme
package is used to obtain estimates for the adjusted differences between
treatment means and the standard error associated with these differences.
Log-transformed BA measures will also be analyzed.
}

\keyword{misc}