\name{description_RepNCAinput}

\alias{description_RepNCAinput}

\title{Description for NCA input data for replicated study}

\description{

Input/Edit Data
   -> subject no.(subj)
   -> sequence (seq)
         ex. 4 periods
           1:Ref.-->Test-->Ref.-->Test
            2:Test-->Ref.-->Test-->Ref.
        or ex. 3 periods
            1:Ref.-->Test-->Test
            2:Test-->Ref.-->Ref.
   -> period (prd)
          1: 1st-treatment period
          2: 2nd-treatment period
          3: 3rd-treatment period
          4: 4th-treatment period
   -> treatment (drug)
          1: Ref.
          2: Test
   -> sampling time
   -> drug plasma/serum/blood concentration (conc)
}

\keyword{misc}