\name{description_size}

\alias{description_size}

\title{Description for sample size input}

\description{

Required data
 1. Theta is the ratio in average BA between the two formulations
    expressed in percentage of the average reference BA.
 2. Theta=Ut/Ur, where Ut and Ur denote the median BA for the Test and
    the Reference products.
 3. CV stands for the intra-subject coefficient of variation.

}

\keyword{misc}