\name{description_TOST_lnCmax}

\alias{description_TOST_lnCmax}

\title{Description for Two One-Sided Test for lnCmax}

\description{
Interpretation:
Ho: MEAN-test/MEAN-ref =< ln(lower acceptance) or  MEAN-test/MEAN-ref >= ln(Upper acceptance)
Ha: ln(lower acceptance) < MEAN-test/MEAN-ref < ln(Upper acceptance)
  If at least one of P value is more than 0.05, we will not reject the null
hypothesis (Ho). 
}
\references{

1. Chow SC and Liu JP. Design and Analysis of Bioavailability-
   Bioequivalence Studies. 3rd ed., Chapman & Hall/CRC, New York (2009).
2. Schuirmann DJ. On hypothesis testing to determine if the mean of a
   normal distribution is continued in a known interval.Biometrics, 37,
   617(1981).
3. Schuirmann DJ. A comparison of the two one-sided tests procedure and the
   power approach for assessing the equivalence of average bioavailability.
   Journal of Pharmacokinetics and Biopharmaceutics, 15, 657-680 (1987).
4. Anderson S and Hauck WW.  A new procedure for testing equivalence in
   comparative bioavailability and other clinical trials. Communications
   in Statistics-Theory and Methods, 12, 2663-2692 (1983).
}

\keyword{misc}