\name{GLManalyze}

\alias{GLManalyze}

\title{GLM analyze function}

\description{
 A two-treatment, two-period, two-sequence (2x2) randomized crossover design,
 the statistical model includes factors accounting for the following sources of
 variation: sequence, subjects nested in sequences, period and treatment.
 GLM was used to obtain estimates for the adjusted differences between treatment
  means and the standard error associated with these differences.
  Log-transformed BA measures were also analyzed.
}
\references{
 Guidance for Industry. Statistical approaches to establishing Bioequivalence. 
}
\keyword{misc}