\encoding{UTF-8}

\name{go2menu}

\alias{go2menu}

\title{List of bear Menu}

\description{
  You can use the functions as follows:
   1. Single dose study
   2. Multiple dose study
   3. *Single dose study   (with ODA)
   4. *Multiple dose study (with ODA)
   5. Quit
}

\keyword{misc}