\encoding{UTF-8}

\name{helper.func}

\alias{helper.func}

\title{helper.func}

\description{
some helper functions from: Cookbook for R, 
http://www.cookbook-r.com/Graphs/Plotting_means_and_error_bars_%28ggplot2%29/;
used to help ggplot()
}

\author{Winston Chang}

\keyword{summarySE, normDataWithin, summarySEwithin}