\name{logdata}

\alias{logdata}

\title{Sample size estimation for log transformation data}

\description{
  This function will help you to choose appropriate sample size.
}
\references{
  Hauschke D, Steinijans VW, Diletti E and Burke M.  Sample size determination
  for bioequivalence assessment using a multiplicative model.  J. Pharmacokin.
  Biopharm. 20:557-561 (1992).
}

\keyword{misc}