\encoding{UTF-8}

\name{MultipleBANOVAdata}

\alias{MultipleBANOVAdata}

\title{Input/Edit data  for ANOVA function for multiple dose}

\description{
  ->subject no.(subj)
  ->drug
       1:Reference
       2:Test
  ->sequence (seq)
       Sequence 1:Reference-->Test sequence
       Sequence 2:Test-->Reference sequence
   ->period (prd)
       Period 1: first treatmetn period
       Period 2: second treatmetn period
   ->Css_max
   ->AUCss(tau): area under the predicted plasma concentration time curve (time = tau)
   ->LnCss_max: Log-transformed Cmax
   ->LnAUC0ss(tau): Log-transformed AUC0ss(tau)
}

\keyword{misc}