\encoding{UTF-8}

\name{MultipleBANOVAmenu}

\alias{MultipleBANOVAmenu}

\title{List of ANOVA Menu for multiple dose}

\description{
  You can use the functions as follows:
   1.Statistical analysis (ANOVA(lm), 90CI...)
   2.Demo for Statistical analysis (ANOVA(lm), 90CI...)
}

\keyword{misc}