\encoding{UTF-8}

\name{MultipleBANOVAoutput}

\alias{MultipleBANOVAoutput}

\title{Output of ANOVA function for multiple dose}

\description{
  We provides several txt outputs.
  1.ANOVA stat.txt
    -->ANOVA:Css_max, AUCss(tau), lnCss_max, lnAUCss(tau)
    -->90CI: lnCss_max, lnAUCss(tau)
}

\keyword{misc}