\encoding{UTF-8}

\name{MultipleBANOVAplot}

\alias{MultipleBANOVAplot}

\title{BANOVAplot for multiple dose}


\description{
  We provides several pdf. outputs about BANOVA plots for outlier detections.
    1.Normal Probability Plot of lnCss_max (intrasubj)
    2.Normal Probability Plot of lnCss_max (intersubj)
    3.lnCss_max(expected value) vs. studentized residuals(intrasubj)
    4.lnCss_max(expected value) vs. studentized residuals(intersubj)
    5.Normal Probability Plot of lnAUCss(tau) (intrasubj)
    6.Normal Probability Plot of lnAUCss(tau) (intersubj)
    7.lnAUCss(tau)(expected value) vs. studentized residuals(intrasubj)
    8.lnAUCss(tau)(expected value) vs. studentized residuals(intersubj)
}

\keyword{misc}