\encoding{UTF-8}

\name{Multipledemomenu}

\alias{Multipledemomenu}

\title{menu for NCA demo file}

\description{
lambda z est. from the exact 3 data points,
lambda z est. with adjusted R sq. (ARS),
lambda z est. with Akaike information criterion (AIC),
lambda z est. with Two-Times-Tmax method (TTT),
lambda z est. with TTT and ARS,
lambda z est. with TTT and AIC
}

\keyword{misc}