\encoding{UTF-8}

\name{MultipleNCAoutput}

\alias{MultipleNCAoutput}

\title{Output of NCA for multiple dose}

\description{
We provides several txt outputs.
1. NCA PK.txt:
     -->Css_max, Tss_max, AUCss(tau), ln(Css_max), ln(AUCss(tau)),
        MRTss(tau), T1/2(z), Vd/F, lambda, Clss/F, Cav and Fluctuation
     --> PK parameter summary (NCA outputs)
2. plots.pdf: individual linear and semilog plots, spaghetti plots, and
         mean conc. plots.

3. Statistical summaries.txt
}

\keyword{misc}