\encoding{UTF-8}

\name{MultipleParaAICoutput}

\alias{MultipleParaAICoutput}

\title{Output for Adjusted R squared (ARS) method for parallel with multiple dose study}

\description{
We provides several txt outputs.
1. NCA PK.txt:
    -->Css_max, Tss_max, AUCss(tau), ln(Css_max), ln(AUCss(tau)),
        MRTss(tau), T1/2(z), Vd/F, lambda, Clss/F, Cav and Fluctuation
     --> PK parameter summary (NCA outputs)
2. plots.pdf: individual linear and semilog plots, spaghetti plots, and
         mean conc. plots.

3. Statistical summaries.txt
}

\keyword{misc}