\encoding{UTF-8}

\name{MultipleParademoMIX}

\alias{MultipleParademoMIX}

\title{Statistical analysis (lme, 90CI...)for demo file for parallel with multiple dose study}

\description{
Statistical analysis (lme, 90CI...):
With a two-treatment, two-sequence, one-period, parallel design, bear
deploys the linear mixed effect model (lme).  The statistical model
includes a factor regarding only one source of variation - treatment.
There are no sources of variation associated with sequence or period
because there are no sequences or periods in a parallel design.  Moreover,
lme in nlme package is used to obtain estimates for the adjusted
differences between treatment means and the standard error.
Log-transformed BA measures will also be analyzed.
}

\keyword{misc}