\encoding{UTF-8}

\name{MultipleParaMIXanalyze}

\alias{MultipleParaMIXanalyze}

\title{Split data to perform lme function for parallel with multiple dose study}

\description{
 Split data for lme function
}
\references{
 Guidance for Industry. Statistical approaches to establishing Bioequivalence.
}
\keyword{misc}