\encoding{UTF-8}

\name{MultipleParaMIXdata}

\alias{MultipleParaMIXdata}

\title{Input/Edit data  for lme function for parallel with multiple dose study}

\description{
  ->subject no.(subj)
  ->drug
       1:Reference
       2:Test
   ->Css_max
   ->AUCss(tau): area under the predicted plasma concentration time curve (time = tau)
   ->LnCss_max: Log-transformed Cmax
   ->LnAUC0ss(tau): Log-transformed AUC0ss(tau)

}

\keyword{misc}