\encoding{UTF-8}

\name{MultipleParaMIXoutput}

\alias{MultipleParaMIXoutput}

\title{Output of lme function for multiple dose}

\description{
  We provides several txt outputs.
  1.lme stat.txt
    -->lme: Css_max, AUCss(tau), lnCss_max, lnAUCss(tau)
    -->90CI: lnCss_max, lnAUCss(tau)
}

\keyword{misc}