\encoding{UTF-8}

\name{MultipleParaNCA.MIXdata}

\alias{MultipleParaNCA.MIXdata}

\title{Input/Edit data  for NCA and lme function for parallel with multiple dose study}

\description{
Input/Edit Data
  ->subject no.(subj)
   ->drug
     ref.: 1
     test: 2
   ->time
   ->concentration (conc)
}

\keyword{misc}