\encoding{UTF-8}

\name{MultipleParaNCAdemo}

\alias{MultipleParaNCAdemo}

\title{Select the exact 3 data points(NCA) for demo function for parallel study}

\description{
Noncompartmental analysis (NCA) approach is used to compute AUCs and
the terminal elimination rate constants (lambda z) for drug plasma
concentration.  The linear trapezoidal method is applied to calculate
AUC(time 0 to the last measurable Cp).  The extrapolated AUC (from time
of the last measurable Cp to time infinity) is equal to the last measurable
Cp divided by lambda z.
}

\keyword{misc}