\encoding{UTF-8}

\name{MultipleParaNCAplot}

\alias{MultipleParaNCAplot}

\title{NCAplot for parallel study}


\description{
  We provides several pdf. outputs about NCA plots.
    1.Conc. vs. Time for individual subjects
    2.Log10(Conc.)  vs. Time for individual subjects
    3.Test (.Conc. vs. Time) for all subjects
    4.Reference (.Conc. vs. Time) for all subjects
    5.Conc.(mean+-sd) vs Time for test and reference
}

\keyword{misc}