\encoding{UTF-8}

\name{Multiplestat1menu}

\alias{Multiplestat1menu}

\title{List of Statistical analysis non-replicated and replicated study}

\description{
  You can use the functions as follows:
   1. Statistical analysis for non-replicated crossover study,
   2. Statistical analysis for replicated crossover study,
}

\keyword{misc}