\encoding{UTF-8}

\name{Multiplestatmenu}

\alias{Multiplestatmenu}

\title{List of NCA-->statistics non-replicated and replicated study}

\description{
  You can use the functions as follows:
   1. NCA for non-replicated crossover study,
   2. NCA for replicated crossover study,
}

\keyword{misc}