\name{NCA.BANOVAdata}

\alias{NCA.BANOVAdata}

\title{Input/Edit data  for NCA and ANOVA function}

\description{
Input/Edit Data
  ->subject no.(subj)
  ->sequence (seq)
       Sequence 1:Reference-->Test sequence
       Sequence 2:Test-->Reference sequence
   ->period (prd)
       Period 1: first treatmetn period
       Period 2: second treatmetn period
   ->time
   ->concentration (conc)
}

\keyword{misc}