\name{NCA.BANOVAmenu}

\alias{NCA.BANOVAmenu}

\title{List of NCA and ANOVA Menu}

\description{
  You can use the functions as follows:
   1.NCA-->ANOVA
   2.Run demo for NCA (exact 3 data points)--> Statistical analysis
   3.Run demo for NCA (ARS)--> Statistical analysis
   4.Run demo for NCA (TTT)--> Statistical analysis
   
}

\keyword{misc}