\name{NCAoutput}

\alias{NCAoutput}

\title{Output of NCA}

\description{
  We provides several txt outputs.
    1.NCA PK:
      -->NCA all: Cmax, Tmax, AUC0t, AUC0inf, AUC0t/AUC0inf, ln(Cmax),
                  ln(AUC0t),ln(AUC0inf), MRT0inf, T1/2, Vd/F, lambda, Cl/F
      -->PK summaries
    2.Statistical summaries
}

\keyword{misc}