\encoding{UTF-8}

\name{NCAreglplot}

\alias{NCAreglplot}

\title{Plot regression lines and save all these plots in a .pdf in NCA}

\description{
This function is for plotting regression lines of data manual slection for lambda_z,
from previously saved data point selection file (.RData).
}

\keyword{misc}