\encoding{UTF-8}

\name{normDataWithin}

\alias{normDataWithin}

\title{normDataWithin}

\description{
normDataWithin():- helper functions from: Cookbook for R, 
http://www.cookbook-r.com/Graphs/Plotting_means_and_error_bars_%28ggplot2%29/;
used to help ggplot()
}

\author{Winston Chang}

\keyword{summarySE, normDataWithin, summarySEwithin}