\name{Paradata}

\alias{Paradata}

\title{Sample size estimation for log transformation data for parallel study}

\description{
  This function will help you to choose appropriate sample size.
}
\references{
  1. Hauschke D, Steinijans VW, Diletti E and Burke M.  Sample size determination
     for bioequivalence assessment using a multiplicative model.  J. Pharmacokin.
     Biopharm. 20:557-561 (1992).
  2. Julious SA.  Tutorial in biostatistics: Sample sizes for clinical trials
     with normal data.  Statist. Med. 23:1921-1986 (2004)
}

\keyword{misc}