\encoding{UTF-8}

\name{ParaMIXdata}

\alias{ParaMIXdata}

\title{Input/Edit data  for lme function for parallel study}

\description{
  ->subject no.(subj)
  ->drug
       1:Reference
       2:Test
   ->Cmax
   ->AUC0t: area under the predicted plasma concentration time curve for
            test data. (time = 0 to t)
   ->AUC0INF: area under the predicted plasma concentration time curve for
             test data. (time = 0 to infinity)
   ->LnCmax: Log-transformed Cmax
   ->LnAUC0t: Log-transformed AUC0t
   ->LnAUC0INF: Log-transformed AUC0INF

}

\keyword{misc}