\encoding{UTF-8}

\name{ParaNCAanalyze}

\alias{ParaNCAanalyze}

\title{NCA analyze function for parallel study}

\description{
We provide noncompartmental analysis (NCA) approach to compute AUCs and terminal
elimination rate constant (kel) for plasma concentration.  Here we provide
six methods, exact 3 data points, ARS, TTT, AIC, TTT and ARS, and TTT and AIC.
}

\keyword{misc}