\name{ParaTTToutput}

\alias{ParaTTToutput}

\title{Output for  Two Times Tmax (TTT) method for parallel study}

\description{
We provides several txt outputs.
1. NCA PK.txt:
     --> Cmax, Tmax, AUC0t, AUC0inf, AUC0t/AUC0inf, ln(Cmax), ln(AUC0t),
         ln(AUC0inf), MRT0inf, T1/2, Vd/F, lambda, and Cl/F
     --> PK parameter summary (NCA outputs)
2. plots.pdf: individual linear and semilog plots, spaghetti plots, and
         mean conc. plots.

3. Statistical summaries.txt
}

\keyword{misc}