\name{rawdata}

\alias{rawdata}

\title{Sample size estimation for raw data}

\description{
  This function will help you to choose appropriate sample size.
}
\references{
  Liu JP, Show SC.  Sample size determination for the tqo one-sided tests
  procedure in bioequivalence.  J. Pharmacokin. Biopharm. 20:101-104 (1992).
}

\keyword{misc}