\encoding{UTF-8}

\name{RepNCA.MIXdata}

\alias{RepNCA.MIXdata}

\title{Input/Edit data  for NCA and lme function for replicated study}

\description{
Input/Edit Data
  ->subject no.(subj)
  ->sequence (seq)
   ->period (prd)
       Period 1: first treatmetn period
       Period 2: second treatmetn period
   ->drug
   ->time
   ->concentration (conc)
}

\keyword{misc}