\encoding{UTF-8}

\name{RepNCAdata}

\alias{RepNCAdata}

\title{Input/Edit data  for NCA function for replicated study}

\description{
Input/Edit Data
  ->subject no.(subj)
  ->sequence (seq)
       Sequence 1:Reference-->Test sequence
       Sequence 2:Test-->Reference sequence
   ->period (prd)
       Period 1: first treatmetn period
       Period 2: second treatmetn period
   ->drug
   ->time
   ->concentration (conc)
}

\keyword{misc}