\name{sizemenu}

\alias{sizemenu}

\title{List of Sample size estimation Menu}

\description{
  You can use the functions as follows:
   1.Sample size estimation (Raw data)
   2.Sample size estimation (Log transformation)
   3.Demo for sample size estimation (Raw data)
  }

\keyword{misc}