\encoding{UTF-8}

\name{BANOVA}

\alias{BANOVA}

\title{Statistical analysis (ANOVA(lm), 90CI...)}

\description{
With a two-treatment, two-period, two-sequence randomized crossover
design, ANOVA statistical model includes factors of the following sources:
sequence, subjects nested in sequences, period and treatment. ANOVA will
be applied to obtain estimates for the adjusted differences between
treatment means and the standard error associated with these differences.
Log-transformed BA measures will also be analyzed.

We investigate the assessment of equivalence in intra-subject variabilities 
of bioavailability between formulations.  Point and interval estimates for the 
inter-subject and intra-subject variabilities are also provided.

Moreover, a normal distribution test procedure based on Spearman's rank and  
Pearson's correlation coefficient and Pitman-Morgan's adjusted F test are provided.

Finally, several tests for assumptions using the inter-subject and intra-subject 
residuals are discussed.  Statistical tests for detection of outlying subjects 
such as Hotelling T2 are presented.
      
}
\references{
1. Chow SC and Liu JP. Design and analysis of bioavailability-
   bioequivalence studies. Chapman & Hall/CRC, New York (2009). 
2. Liu JP and Weng CS.  Detection of outlying data in bioavailability-
   bioequivalence studies. Statist. Med. 10, 1375-1389 (1991).                                                        
}
\keyword{misc}