\name{description_plot}

\alias{description_plot}

\title{Description for plot}

\description{

Authors: Hsin-ya Lee, Yung-jin Lee
100, Shih-chuan 1st Rd.
College of Pharmacy,
Kaohsiung Medical University,
Kaoshiung, Taiwan 80708
E-mail: hsinyalee@gmail.com,
        pkpd.taiwan@gmail.com
bear's website: http://pkpd.kmu.edu.tw/bear
R website: www.r-project.org
}

\keyword{misc}