\name{logdata}

\alias{logdata}

\title{Sample size estimation for log transformation data for crossover and replicated study}

\description{
  This function will help you to choose appropriate sample size.
}
\references{
  R-code based on SAS-code by
  (1) B. Jones and M.G. Kenward
      Design and Analysis of Cross-Over Trials
      Chapman & Hall/CRC, Boca Raton (2nd Edition 2000)
  (2) S. Patterson and B. Jones
      Bioequivalence and Statistics in Clinical Pharmacology
      Chapman & Hall/CRC, Boca Raton (2006)
 /*** WARNING : PROGRAM OFFERED FOR USE WITHOUT ANY GUARANTEES    ***/
 /*** NO LIABILITY IS ACCEPTED FOR ANY LOSS RESULTING FROM USE OF ***/
 /*** THIS SET OF SAS INTRUCTIONS                                 ***/
      Modification of degrees of freedom according to a
      personal message by D. Hauschke (E-mail 2006-01-05)
      Tested in R-versions 2.6.2 / 2.5.1 / 1.9.1 / 1.9.0
      2008-04-04
  Sample size R-code was referred from Helmut Schuetz
      BEBAC - Consultancy services for Bioequivalence
      and Bioavailability Studies
      1070 Vienna, Austria
  
  Hauschke D, Steinijans VW, Diletti E and Burke M.  Sample size determination
  for bioequivalence assessment using a multiplicative model.  J. Pharmacokin.
  Biopharm. 20:557-561 (1992).
}

\keyword{misc}