\name{NCAanalyze}

\alias{NCAanalyze}

\title{NCA analyze function}

\description{
We provide noncompartmental analysis (NCA) approach to compute AUCs and terminal
elimination rate constant (kel) for plasma concentration.  Here we provide
three methods, exact 3 data points, ARS and TTT.
}

\keyword{misc}